requirements.txt
clip==1.0
matplotlib==3.8.2
numpy==1.25.1
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
Pillow==8.3.2
Pillow==10.3.0
PyYAML==6.0
PyYAML==6.0.1
streamlit==1.33.0
streamlit_authenticator==0.3.2
torch==2.0.1
